requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.1.0
Django==4.1
django-countries==7.3.2
django-dotenv==1.4.2
django-filter==22.1
django-widget-tweaks==1.4.12
harperdb==0.4.1
idna==3.3
linkedin-api==1.1.0
Pillow==9.2.0
requests==2.28.1
sqlparse==0.4.2
typing_extensions==4.3.0
tzdata==2022.1
urllib3==1.26.11
